+++ include/exadg/solvers_and_preconditioners/multigrid/multigrid_preconditioner_base.cpp
@@ -37,7 +37,7 @@
 #include <exadg/solvers_and_preconditioners/multigrid/smoothers/gmres_smoother.h>
 #include <exadg/solvers_and_preconditioners/multigrid/smoothers/jacobi_smoother.h>
 #include <exadg/solvers_and_preconditioners/multigrid/transfer/mg_transfer_global_coarsening.h>
-#include <exadg/solvers_and_preconditioners/multigrid/transfer/mg_transfer_mg_level_object.h>
+#include <exadg/solvers_and_preconditioners/multigrid/transfer/mg_transfer_global_refinement.h>
 #include <exadg/solvers_and_preconditioners/util/compute_eigenvalues.h>
 
 namespace ExaDG
