+++ include/exadg/solvers_and_preconditioners/multigrid/transfer/mg_transfer_global_refinement.cpp
@@ -19,10 +19,9 @@
  *  ______________________________________________________________________
  */
 
-#include <exadg/solvers_and_preconditioners/multigrid/transfer/mg_transfer_mg_level_object.h>
-
 #include <exadg/solvers_and_preconditioners/multigrid/levels_hybrid_multigrid.h>
 #include <exadg/solvers_and_preconditioners/multigrid/transfer/mg_transfer_c.h>
+#include <exadg/solvers_and_preconditioners/multigrid/transfer/mg_transfer_global_refinement.h>
 #include <exadg/solvers_and_preconditioners/multigrid/transfer/mg_transfer_h.h>
 #include <exadg/solvers_and_preconditioners/multigrid/transfer/mg_transfer_p.h>
